[Refactor] views 패키지 구조 변경 - 라우터에 등록한 Vue파일을 기준으로 패키지 생성. 해당 Vue파일에서 참조하고 있는 컴포넌트는 같은 패키지에서 관리
