+++ src/views/game/joinTeamSelection/GameJoinTeamSelectionPage.vue
