+++ src/views/myTeam/detail/button/GameRecordPageMoveBtn.vue
